refactor: from core to video module with its generic class for depth and other kind of videos
